fix : Fix Playing Video in Project Activity
Fix frameCount from intent.
Fix playing when video is finished.

+++ app/src/main/java/com/teamisland/zzazz/utils/LoadingDialog.kt
@@ -181,7 +181,7 @@ class LoadingDialog(context: Context, private val request: Int) :
                         putExtra(TrimmingActivity.IMAGE_PATH, originPath)
                         putExtra(
                             TrimmingActivity.VIDEO_FRAME_COUNT,
-                            dataBinder.rangeExclusiveEndIndex - dataBinder.rangeStartIndex
+                            (dataBinder.rangeExclusiveEndIndex - dataBinder.rangeStartIndex).toInt()
                         )
                     }.also { startActivity(context, it, null) }
